Move checkIDP
